+++ ichnaea/api/transfer/__init__.py
@@ -1,4 +1,0 @@
-"""
-Implementation of the public HTTP APIs related to transferring
-aggregated station data from one service instance to another.
-"""